Remove Addressbook module from MarginCalculator (#338)
* refactor

* fix tests

+++ specs/harness/OtokenHarnessA.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.6.0;
-import "../../contracts/Otoken.sol";
+import '../../contracts/Otoken.sol';
 
 contract OtokenHarnessA is Otoken {
-    function havocTotalSupply(uint newVal) public {
-        _totalSupply = newVal;
-    }
+  function havocTotalSupply(uint256 newVal) public {
+    _totalSupply = newVal;
+  }
 }
